requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.2.0
Django==4.2.4
djangorestframework==3.15.2
djangorestframework-api-key==3.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.5.1
packaging==24.1
Pillow==10.0.1
prompt_toolkit==3.0.50
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
